+++ src/Fable.Transforms/Replacements.fs
@@ -566,6 +566,12 @@ let rec equals (com: ICompiler) ctx r equal (left: Expr) (right: Expr) =
     | DeclaredType _ ->
         Helper.LibCall(com, "Util", "equals", Boolean, [ left; right ], ?loc = r)
         |> is equal
+    | Array(_, ResizeArray) ->
+        // ResizeArray (System.Collections.Generic.List) uses reference equality in .NET, see #3718
+        if equal then
+            makeBinOp r Boolean left right BinaryEqual
+        else
+            makeBinOp r Boolean left right BinaryUnequal
     | Array(t, _) ->
         let f = makeEqualityFunction com ctx t